Fix storage size handling in Node registration and file upload processes
- Convert storage values from gigabytes to bytes during Node registration.
- Ensure all storage-related calculations are done in bytes.
- Update file upload process to correctly check available storage in bytes.
- Prevent file upload if there is insufficient available storage on the Node.

+++ DecentraCloud/DecentraCloud.API/Models/Node.cs
@@ -10,7 +10,7 @@ namespace DecentraCloud.API.Models
         [BsonRepresentation(BsonType.ObjectId)]
         public string Id { get; set; }
         public string UserId { get; set; }
-        public int Storage { get; set; }
+        public long Storage { get; set; }
 
         [BsonElement("allocatedFileStorage")]
         public StorageStats AllocatedFileStorage { get; set; } = new StorageStats();
